+++ RedPandaIDE/settings.cpp
@@ -1365,7 +1365,7 @@ void Settings::Editor::doSave()
     saveValue("gutter_visible", mGutterVisible);
     saveValue("gutter_auto_size", mGutterAutoSize);
     saveValue("gutter_left_offset",mGutterLeftOffset);
-    saveValue("gutter_right_offset",mGutterRightOffset);
+    saveValue("gutter_right_offset2",mGutterRightOffset);
     saveValue("gutter_digits_count", mGutterDigitsCount);
     saveValue("gutter_show_line_numbers",mGutterShowLineNumbers);
     saveValue("gutter_add_leading_zero",mGutterAddLeadingZero);
@@ -1514,7 +1514,11 @@ void Settings::Editor::doLoad()
     mGutterVisible = boolValue("gutter_visible",true);
     mGutterAutoSize = boolValue("gutter_auto_size",true);
     mGutterLeftOffset = intValue("gutter_left_offset",6);
-    mGutterRightOffset = intValue("gutter_right_offset",24);
+    mGutterRightOffset = intValue("gutter_right_offset",-1);
+    if (mGutterRightOffset>0)
+        mGutterRightOffset = std::max(0, mGutterRightOffset-20);
+    else
+        mGutterRightOffset = intValue("gutter_right_offset2",4);
     mGutterDigitsCount = intValue("gutter_digits_count",1);
     mGutterShowLineNumbers = boolValue("gutter_show_line_numbers",true);
     mGutterAddLeadingZero = boolValue("gutter_add_leading_zero",false);