+++ NEWS.md
@@ -55,6 +55,7 @@ Red Panda C++ Version 2.27
   - enhancement: Make output in the debug console cleaner.
   - enhancement: Execute the last debug command in the debug console if ENTER pressed.
   - change: When debugging, don't auto set focus to the editor.
+  - enhancement: Folding button scales with editor font.
 
 Red Panda C++ Version 2.26
   - enhancement: Code suggestion for embedded std::vectors.